requirements.txt
azure-functions
sqlalchemy
pyodbc